fix: extract classifyScoring to its own module to satisfy react-refresh lint
The Vite react-refresh plugin warns when a *.tsx file exports both a
component and a non-component binding, since fast-refresh can't reload
the module cleanly. Moves classifyScoring + ScoringFamily into a sibling
.ts module and updates the chart component plus test to import from
there.
