+++ frontend/src/components/dashboard/ScoreTrendChart.tsx
@@ -27,9 +27,9 @@ import {
 import { useTranslation } from 'react-i18next'
 import type { TrendSeriesPoint, MeasureThreshold } from '../../types'
 import { CHART_COLORS } from '../../constants/chartColors'
+import { classifyScoring, type ScoringFamily } from './scoringFamily'
 
 type ChartMode = 'multiples' | 'overlay'
-type ScoringFamily = 'proportion' | 'continuousVariable' | 'cohort'
 
 interface ScoreTrendChartProps {
   data: TrendSeriesPoint[]
@@ -43,20 +43,6 @@ interface ChartDataPoint {
 }
 
 const SMALL_CHART_HEIGHT = 160
-
-/**
- * PAT-124: classify scoring type into one of three families that share a Y-axis
- * convention. Proportion / ratio / composite all live on a 0–100% scale and are
- * safe to overlay; continuous-variable scores are raw clinical values (HbA1c =
- * 5.6 mmol/L) and must each get an auto-scaled axis with the right unit; cohort
- * is a patient count, not a score, so we render it separately too.
- */
-export function classifyScoring(scoringType: string | undefined): ScoringFamily {
-  if (!scoringType) return 'proportion'
-  if (scoringType === 'continuous-variable') return 'continuousVariable'
-  if (scoringType === 'cohort') return 'cohort'
-  return 'proportion'
-}
 
 function buildChartRows(points: TrendSeriesPoint[]): {
   rows: ChartDataPoint[]
